Some work on JobPostings and app.jsx
